feat: will reject jobs, which attempt setting job names by 'slurm_extra' (#93)
Occasionally, users reported that their attempt to overwrite SLURM job
names leads to faulty behaviour (job status does not get queried
properly, and a workflow will stall). This is understandable from the
code and documented accordingly.

The PR checks for the job status using a regex - simple string matching
might be too error-prone (e.g. "-J" can be part of a wildcard).

The decision is here, that the workflow will abort (raising
WorkflowError and explaining and pointing to the docs).

Please check whether this is "likeable".

I do not know how to implement a test. The other executors leave me
clueless.

---------

Co-authored-by: Filipe G. Vieira <1151762+fgvieira@users.noreply.github.com>