fix: sbatch stderr parsing (#161)
will hopefully fix #157 

The issue is, that submission joined `stderr` and `stdout` of the
`sbatch` call. Without add-ons `sbatch` only emits to `stdout` and to
`stderr` only in the case of an error. However, admins can add
informative messages to `stderr`, when this occurs, parsing the message
for the JobID failed. Now, `stderr` and `stdout` are considered
separately.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Enhanced error handling during SLURM job submission, providing clearer
feedback on failures.
	- Improved job ID retrieval by stripping whitespace from the output.

- **Bug Fixes**
- Addressed issues with job submission failures by capturing both
standard output and error messages.

- **Chores**
- Minor adjustments to logging for better clarity during job submission
and error reporting.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Filipe G. Vieira <1151762+fgvieira@users.noreply.github.com>
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>