feat: Improved Mastodon Bot (#183)
We have a Mastodon release bot. This PR attempts to test a generic
message format. It therefore also splits the action yaml into the action
definition and the release script for better maintenance.

It is an experiment.


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Added an automated Mastodon release notification script for the
Snakemake SLURM executor plugin.
- Updated GitHub Actions workflow to utilize the new script for posting
release updates to Mastodon automatically.

- **Chores**
- Enhanced the release communication process by abstracting the posting
logic into a dedicated shell script for Mastodon notifications.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->