fix: logdir misconstruct when leading slash in wildcard (#220)
This is a minor fix, which addresses issue #214: A slurm logdirectory
will start with `/` if a wildcard starts with `/`. After discussing with
@johanneskoester , allowing for slashes in wildcards should not be an
issue. Hence, the escape in this PR. Also,
https://github.com/snakemake/snakemake/issues/3318 can be considered
solved.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced functionality for requesting GPU resources in SLURM jobs,
with detailed documentation on methods for specifying GPUs and
associated CPU settings.

- **Bug Fixes**
- Improved handling of job resource specifications to enforce clarity
between GRES and GPU inputs, preventing potential conflicts.

- **Dependencies**
- Updated the version of the `snakemake-executor-plugin-slurm-jobstep`
dependency from `^0.2.0` to `^0.3.0`, which may include enhancements and
bug fixes.

- **Documentation**
- Added a new section on "GPU Jobs" outlining how to request GPU
resources with examples and clarifications on resource specifications.

- **Tests**
- Removed unnecessary import from test files, streamlining the test code
without affecting functionality.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->