feat: adding reservation flag (#323)
This PR adds a `--reservation` flag. Usually, for teaching, we recommend
relying on SLURM's `magnetic` flag. This way, users will not need to
specify a reservation. However, not all reservations can be made
magnetic nor for all courses. Hence, the introduction of a new flag.

Supersedes PR #285 

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Added support for specifying a SLURM reservation when submitting jobs,
allowing users to include a reservation string during job submission.

- **Documentation**
- Added a section explaining how to use SLURM reservations, including
usage recommendations and best practices for teaching or testing
scenarios.
- Included details about the `--slurm-reservation` flag and noted that
reservation validation and eligibility checks are not performed by the
plugin.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Filipe G. Vieira <1151762+fgvieira@users.noreply.github.com>