feat: measuring compute efficiency per job (#221)
The aim of this PR is:

- measure compute efficiency similar to `seff`
- report compute efficiency and memory footprint per job of the workflow
(in a logfile) and
- issue a warning, if it falls below a configurable threshold

Note: `seff` reports so-called "Memory Efficiency". What is meant is:
"Memory Usage", because an application which reserve a compute node to
compute and hardly uses RAM to do so, will have an apparently low
"Memory Efficiency". It needs to reserve the memory of that node, might
be highly efficient, but will not have used memory.

The resulting code of this PR will hence NOT report warnings about
memory usage.

This PR is related to issue #147

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced an option to generate an efficiency report for SLURM jobs,
summarizing resource usage for all workflow jobs.
- Added settings to enable efficiency reporting, specify the report file
location, and set a CPU efficiency threshold for warnings.
- Efficiency reports are automatically generated upon executor shutdown
when enabled.

- **Documentation**
- Added instructions on generating and interpreting job efficiency
reports with the new reporting option.

- **Bug Fixes**
- Improved error messages and logging for log file cleanup and SLURM job
data handling.

- **Tests**
- Added automated tests to verify the creation and content of the
efficiency report.

- **Chores**
- Added pandas and numpy as new dependencies and upgraded Snakemake and
pandas versions in development dependencies.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: David Laehnemann <david.laehnemann@hhu.de>
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>